Bradesco_GenAI_Dados/ML_a_GenAI_ChatGPT_Whisper_Python/requirements.txt
sounddevice
scipy
openai-whisper
python-dotenv
openai
gtts
openai
gtts

# Comando para instalar as dependências listadas neste arquivo:
# pip install -r requirements.txt 